# designdocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "note-taker API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/notes": {
      "get": {
        "summary": "List or search notes",
        "description": "Return a list of notes, optionally filtered or searched.\n\n- **q** \u2014 full-text search query; when provided, `tag` and `reference` are ignored\n- **tag** \u2014 filter by tag (lowercase, e.g. `work`)\n- **reference** \u2014 filter by reference (lowercase, e.g. `alice`)\n\nReturns up to 500 notes ordered by creation date descending.",
        "operationId": "get_notes_notes_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "reference",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reference"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NoteResponse"
                  },
                  "title": "Response Get Notes Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a note",
        "description": "Create a new note from a plain-text body.\n\nTags (`#Tag`) and references (`@Name`) are parsed automatically from the body.\nAdditional tags and references can be injected via the `tags` and `references` fields\nwithout embedding them in the body text.\n\nReturns the created note with its assigned ID and timestamps.",
        "operationId": "create_note_notes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/notes/{note_id}": {
      "get": {
        "summary": "Get a note by ID",
        "description": "Return a single note by its integer ID.\n\nRaises **404** if no note with that ID exists.",
        "operationId": "get_note_by_id_notes__note_id__get",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a note",
        "description": "Replace the body of an existing note and re-parse its tags and references.\n\nThe full body must be supplied \u2014 this is a replace, not a patch.\nRaises **404** if no note with that ID exists.",
        "operationId": "edit_note_notes__note_id__put",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Note Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "description": "Permanently delete a note by ID.\n\nThe deletion is recorded so it can be propagated during sync.\nRaises **404** if no note with that ID exists.",
        "operationId": "remove_note_notes__note_id__delete",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/config": {
      "get": {
        "summary": "Get configuration",
        "description": "Return the current Mnemo configuration, including sync settings and default tags.",
        "operationId": "get_config_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Config Config Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update configuration",
        "description": "Update one or more configuration values. Omitted fields are left unchanged.\n\n- **default_tags** \u2014 tags automatically applied to every new note\n- **sync_adapter** \u2014 `google_drive` or `local_folder`\n- **sync_folder** \u2014 Google Drive folder name (only used when adapter is `google_drive`)\n- **sync_local_path** \u2014 absolute path to the local sync folder (only used when adapter is `local_folder`)\n- **autosync_debounce_ms** \u2014 milliseconds to wait before triggering an autosync after a note change\n\nReturns the full updated configuration.",
        "operationId": "set_config_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Set Config Config Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NoteResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Auto-incrementing integer primary key"
          },
          "uuid": {
            "type": "string",
            "title": "Uuid",
            "description": "Stable UUID used for sync and conflict resolution"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "description": "Full plain-text note body, including any inline tags and references"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Tags parsed from the body (`#Tag`) plus any injected at creation, stored lowercase"
          },
          "references": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "References",
            "description": "References parsed from the body (`@Name`) plus any injected at creation, stored lowercase"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO 8601 UTC timestamp of when the note was created"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO 8601 UTC timestamp of the most recent edit"
          },
          "time_stamp": {
            "type": "string",
            "title": "Time Stamp",
            "description": "ISO 8601 timestamp of the `~{date}` expression in the body, or equal to `created_at` if none was specified"
          }
        },
        "type": "object",
        "required": [
          "id",
          "uuid",
          "body",
          "tags",
          "references",
          "created_at",
          "updated_at",
          "time_stamp"
        ],
        "title": "NoteResponse",
        "example": {
          "body": "Discussed caching strategy with @Alice \u2014 going with Redis. #Backend #Architecture ~{2026-05-18}",
          "created_at": "2026-05-18T09:30:00+00:00",
          "references": [
            "alice"
          ],
          "id": 42,
          "tags": [
            "backend",
            "architecture"
          ],
          "time_stamp": "2026-05-18T00:00:00+00:00",
          "updated_at": "2026-05-18T09:30:00+00:00",
          "uuid": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        }
      },
...
    }
  }
}